+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.0 (January 26, 2016)
+* Enhancement: add [`--negate`](http://sharp.dimens.io/en/stable/api/#negate) support.
+* Enhancement: add [`--threshold`](http://sharp.dimens.io/en/stable/api/#thresholdthreshold) support.
+* Enhancement: update dependencies, including `sharp` (`0.12.x`).
+
 ## 0.2.2 (January 13, 2016)
 * Enhancement: process multiple files in batches of `25`.